Create separate files for the tables
